trac#17068: server updates FAIlastrunClass and FAIlastrunTime after successful FAI runs
In case of a successful FAI-run (installation or softupdate), the server stores information
about the used FAI classes and the used release to the new ldap attribute FAIlastrunClass.
We also store the current timestamp (in epoch-time) to the new ldap attribute FAIlastrunTime
in that case. E.g.:

FAIlastrunClass: ZAK_Basisclient systest1 :wanderer
FAIlastrunTime: 1452679805

This information is required by admins e.g. to know whether all clients (really) have
recieved a planned release-update. At the moment, there is just a ldap attribute
FAIclass which only reflects the planned state and not necessarily the real facts.

The update is done on the distribution server (which has the access rights to change
the corresponding ldap attributes) and not on the client. In order for the server to
know the required information, the server waits until the clmsg_save_fai_log event
and extracts the information from submitted logfiles. This approach has the advantage,
that we don't neeed to change a client installation for this change to get active.

So this change contains:
- a new step postprocess_logfiles that is called after the server has processed
  the clmsg_save_fai_log event
- a new unit-test / system-test for this scenario
- the corresponding ldap.schema changes for this scenario to run in the go-susi
  testsuite