trac#17068: new "Confirmed" symbols in systems-list show if FAIclass matches FAIlastrunClass
If the host has an own FAIclass, it is used. Otherwise we use the FAIclass that
is derived from the object groups.

We compare this FAIclass with the value of FAIlastrunClass and show a symbol in the
column "Release" of the systems-list with the comparison resulti (confirmed-state).

A mouseover text for the new symbol prints details and the value of FAIlastruntTime if present.

Sorting in the coulumn "Release" is still done primary by the release name. For elements
with the same release name, we also sort by the confirmed-state.